Refactor test structure and update import paths.
Moved test files to more appropriate directories for better organization. Adjusted import paths to reflect the new folder structure and improve module resolution clarity.
