+++ workers/main/src/common/utils.test.ts
@@ -4,8 +4,8 @@ vi.mock('../common/../configs', () => ({
   validationResult: { success: true },
 }));
 
-import * as configs from '../common/../configs';
-import { validateEnv } from '../common/utils';
+import * as configs from '../configs';
+import { validateEnv } from './utils';
 
 type ValidationResult = {
   success: boolean;
